Long is available

+++ src/main/java/moze_intel/projecte/api/proxy/IEMCProxy.java
@@ -65,7 +65,7 @@ public interface IEMCProxy
      * @param block The block we want to query
      * @return The block's EMC value, or 0 if there is none
      */
-    int getValue(Block block);
+    long getValue(Block block);
 
     /**
      * Queries the EMC value for the provided item
@@ -74,7 +74,7 @@ public interface IEMCProxy
      * @param item The item we want to query
      * @return The item's EMC value, or 0 if there is none
      */
-    int getValue(Item item);
+    long getValue(Item item);
 
     /**
      * Queries the EMC value for the provided stack
@@ -84,5 +84,5 @@ public interface IEMCProxy
      * @param stack The stack we want to query
      * @return The stack's EMC value, or 0 if there is none
      */
-    int getValue(ItemStack stack);
+    long getValue(ItemStack stack);
 }
