+++ src/main/java/moze_intel/projecte/emc/collector/DumpToFileCollector.java
@@ -10,13 +10,13 @@ import java.io.File;
 import java.io.IOException;
 import java.util.Map;
 
-public class DumpToFileCollector<A extends IValueArithmetic> extends AbstractMappingCollector<NormalizedSimpleStack, Integer, A>
+public class DumpToFileCollector<A extends IValueArithmetic> extends AbstractMappingCollector<NormalizedSimpleStack, Long, A>
 {
 	public static String currentGroupName="default";
 	CustomConversionFile out = new CustomConversionFile();
-	IExtendedMappingCollector<NormalizedSimpleStack, Integer, A> inner;
+	IExtendedMappingCollector<NormalizedSimpleStack, Long, A> inner;
 	final File file;
-	public DumpToFileCollector(File f, IExtendedMappingCollector<NormalizedSimpleStack, Integer, A> inner)
+	public DumpToFileCollector(File f, IExtendedMappingCollector<NormalizedSimpleStack, Long, A> inner)
 	{
 		super(inner.getArithmetic());
 		file = f;
@@ -42,7 +42,7 @@ public class DumpToFileCollector<A extends IValueArithmetic> extends AbstractMap
 	}
 
 	@Override
-	public void setValueBefore(NormalizedSimpleStack something, Integer value)
+	public void setValueBefore(NormalizedSimpleStack something, Long value)
 	{
 		inner.setValueBefore(something, value);
 		if (something == null) return;
@@ -50,7 +50,7 @@ public class DumpToFileCollector<A extends IValueArithmetic> extends AbstractMap
 	}
 
 	@Override
-	public void setValueAfter(NormalizedSimpleStack something, Integer value)
+	public void setValueAfter(NormalizedSimpleStack something, Long value)
 	{
 		inner.setValueAfter(something, value);
 		if (something == null) return;
