+++ src/main/java/moze_intel/projecte/emc/mappers/Chisel2Mapper.java
@@ -20,7 +20,7 @@ import java.util.List;
 
 
 //Thanks to bdew for a first implementation of this: https://github.com/bdew/ProjectE/blob/f1b08624ff47c6cc716576701024cdb38ff3d297/src/main/java/moze_intel/projecte/emc/ChiselMapper.java
-public class Chisel2Mapper implements IEMCMapper<NormalizedSimpleStack, Integer> {
+public class Chisel2Mapper implements IEMCMapper<NormalizedSimpleStack, Long> {
 
 	public final static String[] chiselBlockNames = new String[]{"marble", "limestone", "andesite", "granite", "diorite"};
 
@@ -40,13 +40,13 @@ public class Chisel2Mapper implements IEMCMapper<NormalizedSimpleStack, Integer>
 	}
 
 	@Override
-	public void addMappings(IMappingCollector<NormalizedSimpleStack, Integer> mapper, Configuration config) {
+	public void addMappings(IMappingCollector<NormalizedSimpleStack, Long> mapper, Configuration config) {
 		ICarvingRegistry carvingRegistry = CarvingUtils.getChiselRegistry();
 		if (carvingRegistry == null) return;
 		for (String name: chiselBlockNames) {
 			Block block = Block.getBlockFromName("chisel:" + name);
 			if (block != null) {
-				mapper.setValueBefore(NormalizedSimpleStack.getFor(block), 1);
+				mapper.setValueBefore(NormalizedSimpleStack.getFor(block), 1L);
 			}
 		}
 
@@ -55,7 +55,7 @@ public class Chisel2Mapper implements IEMCMapper<NormalizedSimpleStack, Integer>
 		}
 	}
 
-	protected void handleCarvingGroup(IMappingCollector<NormalizedSimpleStack, Integer> mapper, Configuration config, ICarvingGroup group) {
+	protected void handleCarvingGroup(IMappingCollector<NormalizedSimpleStack, Long> mapper, Configuration config, ICarvingGroup group) {
 		//XXX: Generates way too much Configs
 		/*if (!config.getBoolean(group.getName(), "enableCarvingGroups", true, "Enable ICarvingGroup with name=" + group.getName() + (group.getOreName() == null ? "" :  " and oreName=" + group.getOreName())) ) {
 			return;
