+++ src/main/java/moze_intel/projecte/emc/mappers/customConversions/CustomConversionMapper.java
@@ -31,7 +31,7 @@ import java.io.OutputStream;
 import java.io.Reader;
 import java.util.Map;
 
-public class CustomConversionMapper implements IEMCMapper<NormalizedSimpleStack, Integer>
+public class CustomConversionMapper implements IEMCMapper<NormalizedSimpleStack, Long>
 {
 	public static final ImmutableList<String> defaultfilenames = ImmutableList.of("metals", "example", "ODdefaults");
 
@@ -55,7 +55,7 @@ public class CustomConversionMapper implements IEMCMapper<NormalizedSimpleStack,
 	}
 
 	@Override
-	public void addMappings(IMappingCollector<NormalizedSimpleStack, Integer> mapper, Configuration config)
+	public void addMappings(IMappingCollector<NormalizedSimpleStack, Long> mapper, Configuration config)
 	{
 		File customConversionFolder = getCustomConversionFolder();
 		if (customConversionFolder.isDirectory() || customConversionFolder.mkdir()) {
@@ -89,11 +89,11 @@ public class CustomConversionMapper implements IEMCMapper<NormalizedSimpleStack,
 		return new File(PECore.CONFIG_DIR, "customConversions");
 	}
 
-	public static void addMappingsFromFile(Reader json, IMappingCollector<NormalizedSimpleStack, Integer> mapper) {
+	public static void addMappingsFromFile(Reader json, IMappingCollector<NormalizedSimpleStack, Long> mapper) {
 		addMappingsFromFile(parseJson(json), mapper);
 	}
 
-	public static void addMappingsFromFile(CustomConversionFile file, IMappingCollector<NormalizedSimpleStack, Integer> mapper) {
+	public static void addMappingsFromFile(CustomConversionFile file, IMappingCollector<NormalizedSimpleStack, Long> mapper) {
 		Map<String, NormalizedSimpleStack> fakes = Maps.newHashMap();
 		//TODO implement buffered IMappingCollector to recover from failures
 		for (Map.Entry<String, ConversionGroup> entry : file.groups.entrySet())
@@ -117,7 +117,7 @@ public class CustomConversionMapper implements IEMCMapper<NormalizedSimpleStack,
 			if (file.values != null)
 			{
 				if (file.values.setValueBefore != null) {
-					for (Map.Entry<String, Integer> entry : file.values.setValueBefore.entrySet())
+					for (Map.Entry<String, Long> entry : file.values.setValueBefore.entrySet())
 					{
 						NormalizedSimpleStack something = getNSSfromJsonString(entry.getKey(), fakes);
 						mapper.setValueBefore(something, entry.getValue());
@@ -133,7 +133,7 @@ public class CustomConversionMapper implements IEMCMapper<NormalizedSimpleStack,
 				}
 				if (file.values.setValueAfter != null)
 				{
-					for (Map.Entry<String, Integer> entry : file.values.setValueAfter.entrySet())
+					for (Map.Entry<String, Long> entry : file.values.setValueAfter.entrySet())
 					{
 						NormalizedSimpleStack something = getNSSfromJsonString(entry.getKey(), fakes);
 						mapper.setValueAfter(something, entry.getValue());
